Changes to camera

+++ src/movable_object.hpp
@@ -25,17 +25,17 @@
 		virtual const glm::mat4 matrix() const;
 
 		virtual void relative_move(const glm::vec3 &move);
-		virtual void relative_rotate(const glm::vec3 &axis, const double &angle);
+		virtual void relative_rotate(const glm::vec3 &axis, const float &angle);
 
-		virtual void roll(const double angle);
-		virtual void pitch(const double angle);
-		virtual void yaw(const double angle);
+		virtual void roll(const float angle);
+		virtual void pitch(const float angle);
+		virtual void yaw(const float angle);
 
-		virtual void absolute_rotate(const glm::vec3 &axis, const double &angle);
+		virtual void absolute_rotate(const glm::vec3 &axis, const float &angle);
 		virtual void absolute_move(const glm::vec3 &move);
 
 		virtual void set_position(const glm::vec3 &pos);
-		virtual void set_rotation(const glm::vec3 &axis, const double angle);
+		virtual void set_rotation(const glm::vec3 &axis, const float angle);
 
 		virtual const glm::vec3 local_x() const;
 		virtual const glm::vec3 local_y() const;
